gitのプルを強制的に実行するシェルを追加

+++ gitpull.sh
@@ -1,4 +1,26 @@
-git fetch origin
-git reset --hard origin/develop-ikura
-git reset --hard origin/main
-chmod 777 init.sh
+#!/bin/bash
+git branch
+echo "現在のブランチにdevelop-ikuraを強制的に上書きします。"
+echo "よろしいですか？ (y/n)"
+read answer
+
+if [ "$answer" = "y" ]; then
+    git fetch origin
+    git reset --hard origin/develop-ikura
+    chmod 777 init.sh
+else
+    echo "処理を中止しました。"
+    exit
+fi
+
+echo "続けてmomoなどを実行できるように設定します。"
+echo "よろしいですか？ (y/n)"
+read answer
+
+if [ "$answer" = "y" ]; then
+    ./init.sh
+else
+    echo "処理を中止しました。"
+    exit
+fi
+
